Restructure source tree: split macros into macros.rs, tests into tests/
